extracted reular task execution to base class
